renaming app to `Emergency Operation Center`
