+++ android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.eocmobileapp",
+    package = "com.emergencyoperationcenter",
 )
 
 android_resource(
     name = "res",
-    package = "com.eocmobileapp",
+    package = "com.emergencyoperationcenter",
     res = "src/main/res",
 )
