+++ android/app/build.gradle
@@ -129,7 +129,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.eocmobileapp"
+        applicationId "com.emergencyoperationcenter"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 2
