+++ android/app/src/main/java/com/emergencyoperationcenter/MainActivity.java
@@ -1,4 +1,4 @@
-package com.eocmobileapp;
+package com.emergencyoperationcenter;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "eocmobileapp";
+        return "EmergencyOperationCenter";
     }
 }
