+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">eocmobileapp</string>
+    <string name="app_name">Emergency Operation Center</string>
 </resources>
