+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'eocmobileapp'
+rootProject.name = 'EmergencyOperationCenter'
 include ':rn-fetch-blob'
 project(':rn-fetch-blob').projectDir = new File(rootProject.projectDir, '../node_modules/rn-fetch-blob/android')
 include ':react-native-cardview'
