+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "eocmobileapp",
-  "displayName": "eocmobileapp"
+  "name": "EmergencyOperationCenter",
+  "displayName": "Emergency Operation Center"
 }
