+++ ios/EmergencyOperationCenter/Images.xcassets/AppIcon.appiconset/Contents.json
