+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "eocmobileapp",
+  "name": "EmergencyOperationCenter",
   "version": "0.0.1",
   "private": true,
   "scripts": {
